Moved all test setup to doctest_helper
